fix(app): renamed the state folder name
