+++ src/app/app.module.ts
@@ -15,11 +15,11 @@ import { AddPostComponent } from './post/add-post/add-post.component';
 import { EditPostComponent } from './post/edit-post/edit-post.component';
 import { EffectsModule } from '@ngrx/effects';
 import { LoadingComponent } from './shared/components/loading/loading.component';
-import { sharedReducer } from './shared/store/shared.state';
+import { sharedReducer } from './shared/state/shared.state';
 import { AuthEffects } from './auth/state/auth.effects';
 import { AuthTokenIntercepter } from './auth/services/auth-token.intercepter';
 import { StoreRouterConnectingModule } from '@ngrx/router-store';
-import { CustomSerializer } from './shared/store/custom-serializer';
+import { CustomSerializer } from './shared/state/custom-serializer';
 
 @NgModule({
   declarations: [
