+++ src/app/auth/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AuthBlService } from './auth-bl.service';
 import { Store } from '@ngrx/store';
-import { SharedState } from 'src/app/shared/store/shared.state';
+import { SharedState } from 'src/app/shared/state/shared.state';
 import { autoLogoutAction } from '../state/auth.actions';
 import { UserModel } from 'src/app/shared/misc/shared.model';
