+++ src/app/post/post-list/post-list.component.ts
@@ -3,7 +3,7 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { getAllPosts } from '../state/post.selectors';
 import { deletePostAction, loadPostsAction } from '../state/post.action';
-import { SharedState } from 'src/app/shared/store/shared.state';
+import { SharedState } from 'src/app/shared/state/shared.state';
 import { Post } from 'src/app/shared/misc/shared.model';
 
 @Component({
