+++ src/app/post/state/post.selectors.ts
@@ -1,7 +1,7 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { PostState, postsAdapter } from './post.state';
 import { StateEnum } from 'src/app/shared/misc/shared.enum';
-import { getCurrentRoute } from 'src/app/shared/store/shared.selectors';
+import { getCurrentRoute } from 'src/app/shared/state/shared.selectors';
 import { RouterStateUrl } from 'src/app/shared/misc/shared.model';
 
 export const postsSelectors = postsAdapter.getSelectors();
