+++ src/app/shared/state/shared.selectors.ts
